requirements.txt
openai>=1.12.0
requests>=2.31.0
yfinance>=0.2.36
pyyaml>=6.0.1